First works on v2.0
- From now, only GIMP >=2.10 is supported. For old releases, switch to "v1" branch
- Use GResources instead of Pixdata
- Added support to HEIF, XPM and JPEG2000 images (#182, #141, #138 )
- Replaced GIMP_INTERPOLATION_LANCZOS with NOHALO and LOHALO
- Many compiler warnings fixed
- Enhanced Win32 make procedure (needs MSYS2)

+++ src/bimp-serialize.h
@@ -8,11 +8,11 @@ gboolean bimp_deserialize_from_file(gchar*);
 
 gboolean parse_curve_file(
     char*, 
-    int*, guint8**, 
-    int*, guint8**, 
-    int*, guint8**, 
-    int*, guint8**, 
-    int*, guint8**
+    int*, gdouble**, 
+    int*, gdouble**, 
+    int*, gdouble**, 
+    int*, gdouble**, 
+    int*, gdouble**
 );
 
 #endif
